+ Sample amplify now accepts negative values(inverts sample phase) (request 2219) . Fixed a typo in Setup->Colors
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@234 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Ctrl_smp.cpp
@@ -1078,7 +1078,7 @@ void CCtrlSamples::OnNormalize()
 void CCtrlSamples::OnAmplify()
 //----------------------------
 {
-	static UINT snOldAmp = 100;
+	static int16 snOldAmp = 100;
 	SAMPLEVIEWSTATE viewstate;
 	DWORD dwStart, dwEnd;
 	MODINSTRUMENT *pins;
